# Proposal: move from .net standard 2.0 to .net standard 2.1 to allow more memory-efficient solutions