@rasmus 2021-08-30
.NET Standard would also make it possible to use `IAsyncEnumerable` 😄