@rasmus 2021-09-08
.NET Core 2.1 has reached end of life, thus dropping support for it might be a possibility 
https://dotnet.microsoft.com/platform/support/policy/dotnet-core